+++ nest-op-orthorectification/src/main/java/org/esa/nest/gpf/CreateElevationOp.java
@@ -176,7 +176,7 @@ public final class CreateElevationOp extends Operator {
                 final TileGeoreferencing tileGeoRef = new TileGeoreferencing(targetProduct, x0, y0, w, h);
 
                 final float demNoDataValue = dem.getDescriptor().getNoDataValue();
-                final float[][] localDEM = new float[h+2][w+2];
+                final double[][] localDEM = new double[h+2][w+2];
                 DEMFactory.getLocalDEM(dem, demNoDataValue, tileGeoRef, x0, y0, w, h, localDEM);
 
                 final TileIndex trgIndex = new TileIndex(targetTile);
